Preventive check

+++ web/analysis/views.py
@@ -312,7 +312,7 @@ def remove(request, task_id):
                 fs.delete(ObjectId(analysis["network"]["pcap_id"]))
             # Delete dropped.
             for drop in analysis["dropped"]:
-                if results_db.analysis.find({"dropped.object_id": ObjectId(drop["object_id"])}).count() == 1:
+                if "object_id" in drop and results_db.analysis.find({"dropped.object_id": ObjectId(drop["object_id"])}).count() == 1:
                     fs.delete(ObjectId(drop["object_id"]))
             # Delete calls.
             for process in analysis["behavior"]["processes"]:
